+++ Editor/Scripts/CustomEditors/OfflineInputDataEditor.cs
@@ -1,5 +1,5 @@
 using NPTP.InputSystemWrapper.Data;
-using NPTP.InputSystemWrapper.Utilities.Editor;
+using NPTP.InputSystemWrapper.Editor.Utilities;
 using UnityEditor;
 using UnityEngine;