+++ Editor/Scripts/ScriptContentBuilders/ControlSchemeContentBuilder.cs
@@ -1,28 +1,32 @@
-using System.Collections.Generic;
+using NPTP.InputSystemWrapper.Data;
 using NPTP.InputSystemWrapper.Enums;
 using UnityEngine.InputSystem;
 
 namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
 {
-    internal static class ControlSchemeContentBuilder
+    internal class ControlSchemeContentBuilder : ContentBuilder
     {
-        internal static void AddContent(InputActionAsset asset, string markerName, List<string> lines)
+        internal override void AddContent(InputScriptGeneratorMarkerInfo info)
         {
-            switch (markerName)
+            switch (info.MarkerName)
             {
                 case "Members":
-                    foreach (InputControlScheme controlScheme in asset.controlSchemes)
-                        lines.Add($"        {controlScheme.name.AsEnumMember()},");
+                    foreach (InputControlScheme controlScheme in Asset.controlSchemes)
+                        info.NewLines.Add($"        {controlScheme.name.AsEnumMember()},");
                     break;
                 case "EnumToStringSwitch":
-                    foreach (InputControlScheme controlScheme in asset.controlSchemes)
-                        lines.Add($"                {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()} => \"{controlScheme.name}\",");
+                    foreach (InputControlScheme controlScheme in Asset.controlSchemes)
+                        info.NewLines.Add($"                {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()} => \"{controlScheme.name}\",");
                     break;
                 case "StringToEnumSwitch":
-                    foreach (InputControlScheme controlScheme in asset.controlSchemes)
-                        lines.Add($"                \"{controlScheme.name}\" => {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()},");
+                    foreach (InputControlScheme controlScheme in Asset.controlSchemes)
+                        info.NewLines.Add($"                \"{controlScheme.name}\" => {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()},");
                     break;
             }
+        }
+
+        public ControlSchemeContentBuilder(OfflineInputData offlineInputData) : base(offlineInputData)
+        {
         }
     }
 }