+++ Editor/Scripts/ScriptContentBuilders/InputContextContentBuilder.cs
@@ -1,22 +1,22 @@
-using System.Collections.Generic;
 using NPTP.InputSystemWrapper.Data;
-using NPTP.InputSystemWrapper.Utilities.Editor;
-using UnityEngine.InputSystem;
 
 namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
 {
-    internal static class InputContextContentBuilder
+    internal class InputContextContentBuilder : ContentBuilder
     {
-        internal static void AddContent(InputActionAsset asset, string markerName, List<string> lines)
+        internal override void AddContent(InputScriptGeneratorMarkerInfo info)
         {
-            switch (markerName)
+            switch (info.MarkerName)
             {
                 case "Members":
-                    OfflineInputData offlineInputData = EditorAssetGetter.GetFirst<OfflineInputData>();
-                    foreach (InputContextInfo contextInfo in offlineInputData.InputContexts)
-                        lines.Add($"        {contextInfo.Name},");
+                    foreach (InputContextInfo contextInfo in Data.InputContexts)
+                        info.NewLines.Add($"        {contextInfo.Name},");
                     break;
             }
+        }
+
+        public InputContextContentBuilder(OfflineInputData offlineInputData) : base(offlineInputData)
+        {
         }
     }
 }