+++ Editor/Scripts/ScriptContentBuilders/InputPlayerContentBuilder.cs
@@ -7,74 +7,73 @@ using UnityEngine.InputSystem;
 
 namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
 {
-    internal static class InputPlayerContentBuilder
+    internal class InputPlayerContentBuilder : ContentBuilder
     {
-        internal static void AddContent(InputActionAsset asset, string markerName, List<string> lines)
+        internal override void AddContent(InputScriptGeneratorMarkerInfo info)
         {
-            OfflineInputData offlineInputData = Helper.OfflineInputData;
-            switch (markerName)
+            switch (info.MarkerName)
             {
                 case "ControlSchemeEventDefinition":
-                    lines.Add(offlineInputData.EnableMultiplayer
+                    info.NewLines.Add(Data.EnableMultiplayer
                         ? $"        public event Action<{nameof(InputPlayer)}> OnControlSchemeChanged;"
                         : $"        public event Action<{nameof(ControlScheme)}> OnControlSchemeChanged;");
                     break;
                 case "ControlSchemeEventInvocation":
-                    lines.Add(offlineInputData.EnableMultiplayer
+                    info.NewLines.Add(Data.EnableMultiplayer
                     ? "                    OnControlSchemeChanged?.Invoke(this);"
                     : "                    OnControlSchemeChanged?.Invoke(controlScheme);");
                     break;
                 case "ActionsProperties":
-                    foreach (string mapName in Helper.GetMapNames(asset))
-                        lines.Add($"        public {mapName.AsProperty()}Actions {mapName.AsProperty()}" + " { get; }");
+                    foreach (string mapName in Helper.GetMapNames(Asset))
+                        info.NewLines.Add($"        public {mapName.AsProperty()}Actions {mapName.AsProperty()}" + " { get; }");
                     break;
                 case "ActionsInstantiation":
-                    foreach (string map in Helper.GetMapNames(asset))
-                        lines.Add($"            {map.AsProperty()} = new {map.AsType()}Actions(Asset, actionWrapperTable);");
+                    foreach (string map in Helper.GetMapNames(Asset))
+                        info.NewLines.Add($"            {map.AsProperty()} = new {map.AsType()}Actions(Asset, actionWrapperTable);");
                     break;
                 case "EventSystemOptions":
-                    lines.Add($"            uiInputModule.moveRepeatDelay = {offlineInputData.MoveRepeatDelay}f;");
-                    lines.Add($"            uiInputModule.moveRepeatRate = {offlineInputData.MoveRepeatRate}f;");
-                    lines.Add($"            uiInputModule.deselectOnBackgroundClick = {offlineInputData.DeselectOnBackgroundClick.ToString().ToLower()};");
-                    lines.Add($"            uiInputModule.pointerBehavior = UIPointerBehavior.{offlineInputData.PointerBehavior};");
-                    lines.Add($"            uiInputModule.cursorLockBehavior = InputSystemUIInputModule.CursorLockBehavior.{offlineInputData.CursorLockBehavior};");
+                    info.NewLines.Add($"            uiInputModule.moveRepeatDelay = {Data.MoveRepeatDelay}f;");
+                    info.NewLines.Add($"            uiInputModule.moveRepeatRate = {Data.MoveRepeatRate}f;");
+                    info.NewLines.Add($"            uiInputModule.deselectOnBackgroundClick = {Data.DeselectOnBackgroundClick.ToString().ToLower()};");
+                    info.NewLines.Add($"            uiInputModule.pointerBehavior = UIPointerBehavior.{Data.PointerBehavior};");
+                    info.NewLines.Add($"            uiInputModule.cursorLockBehavior = InputSystemUIInputModule.CursorLockBehavior.{Data.CursorLockBehavior};");
                     break;
                 case "PopulateEventSystemActionsPool":
                     HashSet<string> actionIDs = new();
 
                     const string defaultPrefix = "default";
-                    AddEventSystemDefaultAction(offlineInputData.Point, defaultPrefix + nameof(offlineInputData.Point), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.LeftClick, defaultPrefix + nameof(offlineInputData.LeftClick), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.MiddleClick, defaultPrefix + nameof(offlineInputData.MiddleClick), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.RightClick, defaultPrefix + nameof(offlineInputData.RightClick), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.ScrollWheel, defaultPrefix + nameof(offlineInputData.ScrollWheel), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.Move, defaultPrefix + nameof(offlineInputData.Move), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.Submit, defaultPrefix + nameof(offlineInputData.Submit), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.Cancel, defaultPrefix + nameof(offlineInputData.Cancel), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.TrackedDevicePosition, defaultPrefix + nameof(offlineInputData.TrackedDevicePosition), 3, lines, actionIDs);
-                    AddEventSystemDefaultAction(offlineInputData.TrackedDeviceOrientation, defaultPrefix + nameof(offlineInputData.TrackedDeviceOrientation), 3, lines, actionIDs);
+                    AddEventSystemDefaultAction(Data.Point, defaultPrefix + nameof(Data.Point), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.LeftClick, defaultPrefix + nameof(Data.LeftClick), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.MiddleClick, defaultPrefix + nameof(Data.MiddleClick), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.RightClick, defaultPrefix + nameof(Data.RightClick), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.ScrollWheel, defaultPrefix + nameof(Data.ScrollWheel), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.Move, defaultPrefix + nameof(Data.Move), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.Submit, defaultPrefix + nameof(Data.Submit), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.Cancel, defaultPrefix + nameof(Data.Cancel), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.TrackedDevicePosition, defaultPrefix + nameof(Data.TrackedDevicePosition), 3, info.NewLines, actionIDs);
+                    AddEventSystemDefaultAction(Data.TrackedDeviceOrientation, defaultPrefix + nameof(Data.TrackedDeviceOrientation), 3, info.NewLines, actionIDs);
                     
-                    if (offlineInputData.GetEventSystemActionNonNullOverrideCount() == 0) break;
+                    if (Data.GetEventSystemActionNonNullOverrideCount() == 0) break;
                     
-                    lines.Add(string.Empty);
-                    foreach (InputContextInfo inputContextInfo in offlineInputData.InputContexts)
+                    info.NewLines.Add(string.Empty);
+                    foreach (InputContextInfo inputContextInfo in Data.InputContexts)
                         foreach (EventSystemActionSpecification actionOverride in inputContextInfo.EventSystemActionOverrides)
-                            AddEventSystemActionOverride(actionOverride.ActionReference, 3, lines, actionIDs);
+                            AddEventSystemActionOverride(actionOverride.ActionReference, 3, info.NewLines, actionIDs);
                     
                     break;
                 case "DisableAllMapsAndRemoveCallbacksBody":
-                    foreach (string map in Helper.GetMapNames(asset))
-                        lines.Add($"            {map.AsProperty()}.DisableAndUnregisterCallbacks();");
+                    foreach (string map in Helper.GetMapNames(Asset))
+                        info.NewLines.Add($"            {map.AsProperty()}.DisableAndUnregisterCallbacks();");
                     break;
                 case "EnableContextSwitchMembers":
-                    foreach (InputContextInfo contextInfo in offlineInputData.InputContexts)
+                    foreach (InputContextInfo contextInfo in Data.InputContexts)
                     {
-                        lines.Add($"                case {nameof(InputContext)}.{contextInfo.Name}:");
-                        lines.Add($"                    {(contextInfo.EnableKeyboardTextInput ? "Enable" : "Disable")}KeyboardTextInput();");
-                        foreach (string map in Helper.GetMapNames(asset))
+                        info.NewLines.Add($"                case {nameof(InputContext)}.{contextInfo.Name}:");
+                        info.NewLines.Add($"                    {(contextInfo.EnableKeyboardTextInput ? "Enable" : "Disable")}KeyboardTextInput();");
+                        foreach (string map in Helper.GetMapNames(Asset))
                         {
                             bool enable = contextInfo.ActiveMaps.Any(activeMapName => map == activeMapName);
-                            lines.Add($"                    {map.AsProperty()}.{(enable ? "EnableAndRegisterCallbacks" : "DisableAndUnregisterCallbacks")}();");
+                            info.NewLines.Add($"                    {map.AsProperty()}.{(enable ? "EnableAndRegisterCallbacks" : "DisableAndUnregisterCallbacks")}();");
                         }
 
                         foreach (EventSystemActionSpecification spec in contextInfo.EventSystemActionOverrides)
@@ -83,10 +82,10 @@ namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
                             string assignmentValue = spec.ActionReference == null || spec.ActionReference.action == null
                                     ? "null"
                                     : $"eventSystemActionsPool[\"{spec.ActionReference.action.id}\"]";
-                            lines.Add($"                    uiInputModule.{inputModuleActionField} = {assignmentValue};");
+                            info.NewLines.Add($"                    uiInputModule.{inputModuleActionField} = {assignmentValue};");
                         }
                         
-                        lines.Add($"                    break;");
+                        info.NewLines.Add($"                    break;");
                     }
                     break;
             }
@@ -137,6 +136,10 @@ namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
             for (int i = 0; i < tabs; i++) line.Append("    ");
             line.Append($"eventSystemActionsPool.Add(\"{actionId}\", CreateInputActionReferenceToPlayerAsset(\"{actionId}\"));");
             lines.Add(line.ToString());
+        }
+
+        public InputPlayerContentBuilder(OfflineInputData offlineInputData) : base(offlineInputData)
+        {
         }
     }
 }