+++ Editor/Scripts/ScriptContentBuilders/PlayerIDContentBuilder.cs
@@ -1,25 +1,25 @@
-using System.Collections.Generic;
 using NPTP.InputSystemWrapper.Data;
-using NPTP.InputSystemWrapper.Utilities.Editor;
-using UnityEngine.InputSystem;
 
 namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
 {
-    internal static class PlayerIDContentBuilder
+    internal class PlayerIDContentBuilder : ContentBuilder
     {
-        internal static void AddContent(InputActionAsset inputActionAsset, string markerName, List<string> lines)
+        internal override void AddContent(InputScriptGeneratorMarkerInfo info)
         {
-            switch (markerName)
+            switch (info.MarkerName)
             {
                 case "Members":
-                    OfflineInputData offlineInputData = EditorAssetGetter.GetFirst<OfflineInputData>();
-                    int numPlayers = offlineInputData.EnableMultiplayer ? offlineInputData.MaxPlayers : 1;
+                    int numPlayers = Data.EnableMultiplayer ? Data.MaxPlayers : 1;
                     for (int i = 0; i < numPlayers; i++)
                     {
-                        lines.Add($"        Player{i + 1} = {i},");
+                        info.NewLines.Add($"        Player{i + 1} = {i},");
                     }
                     break;
             }
+        }
+
+        public PlayerIDContentBuilder(OfflineInputData offlineInputData) : base(offlineInputData)
+        {
         }
     }
 }