+++ Editor/Scripts/ScriptContentBuilders/RuntimeInputDataContentBuilder.cs
@@ -1,27 +1,31 @@
-using System.Collections.Generic;
 using NPTP.InputSystemWrapper.Bindings;
+using NPTP.InputSystemWrapper.Data;
 using NPTP.InputSystemWrapper.Enums;
 using UnityEngine.InputSystem;
 
 namespace NPTP.InputSystemWrapper.Editor.ScriptContentBuilders
 {
-    internal static class RuntimeInputDataContentBuilder
+    internal class RuntimeInputDataContentBuilder : ContentBuilder
     {
-        internal static void AddContent(InputActionAsset asset, string markerName, List<string> lines)
+        internal override void AddContent(InputScriptGeneratorMarkerInfo info)
         {
-            switch (markerName)
+            switch (info.MarkerName)
             {
                 case "ControlSchemeBindingData":
-                    if (asset.controlSchemes.Count > 0)
-                        lines.Add($"        [Header(\"Input Device Binding Data (Auto-Generated List)\")] [Space]");
-                    foreach (InputControlScheme controlScheme in asset.controlSchemes)
-                        lines.Add($"        [SerializeField] private {nameof(BindingData)} {controlScheme.name.AsField()}BindingData;");
+                    if (Asset.controlSchemes.Count > 0)
+                        info.NewLines.Add($"        [Header(\"Input Device Binding Data (Auto-Generated List)\")] [Space]");
+                    foreach (InputControlScheme controlScheme in Asset.controlSchemes)
+                        info.NewLines.Add($"        [SerializeField] private {nameof(BindingData)} {controlScheme.name.AsField()}BindingData;");
                     break;
                 case "EnumToBindingDataSwitch":
-                    foreach (InputControlScheme controlScheme in asset.controlSchemes)
-                        lines.Add($"                {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()} => {controlScheme.name.AsField()}BindingData,");
+                    foreach (InputControlScheme controlScheme in Asset.controlSchemes)
+                        info.NewLines.Add($"                {nameof(ControlScheme)}.{controlScheme.name.AsEnumMember()} => {controlScheme.name.AsField()}BindingData,");
                     break;
             }
+        }
+
+        public RuntimeInputDataContentBuilder(OfflineInputData offlineInputData) : base(offlineInputData)
+        {
         }
     }
 }