+++ Editor/Scripts/Utilities/EditorScriptGetter.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using NPTP.InputSystemWrapper.Editor;
 using UnityEditor;
 using UnityEngine;
 
-namespace NPTP.InputSystemWrapper.Utilities.Editor
+namespace NPTP.InputSystemWrapper.Editor.Utilities
 {
     internal static class EditorScriptGetter
     {