+++ Runtime/Scripts/Enums/InputContext.cs
@@ -3,6 +3,7 @@ namespace NPTP.InputSystemWrapper.Enums
     public enum InputContext
     {
         // MARKER.Members.Start
+        Default,
         // MARKER.Members.End
     }
 }